transformation of phylo trees to / from command documents. phylo trees use Double instead of BigDecimal for branch lengths.

+++ gluetools-core/src/main/java/uk/ac/gla/cvr/gluetools/core/newick/NewickToPhyloTreeParser.java
@@ -209,7 +209,7 @@ public class NewickToPhyloTreeParser {
 			case NUMBER: 
 				assertTrue(token, colon);
 				assertTrue(token, !number);
-				branch.setLength(new BigDecimal(token.render()));
+				branch.setLength(new BigDecimal(token.render()).doubleValue());
 				number = true;
 				break;
 			case COMMENT:
